requirements.txt
numpy==1.22.3
pandas==1.4.2
pymongo==4.1.1
python-binance==1.0.16
python-dotenv==0.20.0
